Optimize RAM usage: move defaultBands to FLASH, shrink buffers
- defaultBands → const (FLASH): saves ~1792 B from .data to .rodata
- MAX_MESSAGES 8→4: saves ~520 B
- MAX_FILES 20→12: saves ~192 B
- Loot::lastTimeOpen uint32_t→uint16_t: saves ~100 B
- const defaultBands extern declaration in bands.h

+++ src/inc/loot.h
@@ -11,13 +11,13 @@ typedef struct {
   Squelch squelch;
   uint8_t gainIndex : 5;
 
-  uint32_t lastTimeOpen; // TODO: last seen minutes, ++ every min for each
+  uint16_t lastTimeOpen; // minutes since boot (65535 min = 45 days)
   uint16_t duration;
   uint8_t code;
   uint32_t f : 27;
   bool isCd : 1;
   bool open : 1;
-  bool blacklist : 1; // move to flash or freq list
+  bool blacklist : 1;
   bool whitelist : 1;
 } __attribute__((packed)) Loot;
